requirements.txt
Flask
textwrap3
Pillow
PyPDF2
fpdf
